# os/bluestore: avoid the VTABLE-related burden in BitMapAllocator's hotspot